requirements.txt
oauthlib==3.2.1
python-dotenv==1.0.0
requests==2.28.1
requests-oauthlib==1.3.1